+++ AliemMod/AliemPlugin.cs
@@ -56,18 +56,14 @@ public class AliemPlugin : BaseUnityPlugin {
 
         //if (Modules.Config.Debug)
         //    gameObject.AddComponent<TestValueManager>();
-
         Compat.Initialize();
         Tokens.AddTokens(); // register name tokens
         AliemMod.Modules.Language.Init();
         AliemMod.Modules.Language.PrintOutput("aliem.txt");
         ItemDisplays.PopulateDisplays(); // collect item display prefabs for use in our display rules
-
         Buffs.RegisterBuffs(); // add and register custom buffs/debuffs
         Dots.RegisterDots();
-
-        new ContentPacks().Initialize();
-
+        new ContentPacks().Initialize();        
         new AliemMod.Content.Survivors.AliemSurvivor().Initialize();
 
         new AliemPoolManager(transform);